+++ src/domain/map/MapView.tsx
@@ -1,18 +1,12 @@
-import { Component, ReactNode } from "react";
+import { Component } from "react";
 import { withTranslation } from "react-i18next";
 import { Map, TileLayer, ZoomControl } from "react-leaflet";
 
-import OutboundLink from "../../common/a11y/OutboundLink";
 import OSMIcon from "../../common/components/OSMIcon";
-import SMIcon from "../../common/components/SMIcon";
-import FeedbackModal from "../app/AppFeedbackModal";
-import { mobileBreakpoint } from "../app/appConstants";
-import TranslationProvider from "../i18n/I18nTranslationProvider";
 import { View } from "../unit/UnitView";
 import { Unit } from "../unit/unitConstants";
 import { isRetina } from "../utils";
 import Control from "./MapControl";
-import MapDropdownControl from "./MapDropdownControl";
 import MapUnitsOnMap from "./MapUnits";
 import MapUserLocationMarker from "./MapUserLocationMarker";
 import {
@@ -39,10 +33,6 @@ type Props = {
 };
 
 type State = {
-  isMobile: boolean;
-  menuOpen: boolean;
-  aboutModalOpen: boolean;
-  feedbackModalOpen: boolean;
   zoomLevel: number;
 };
 
@@ -50,16 +40,8 @@ class MapView extends Component<Props, State> {
   constructor(props: Props) {
     super(props);
     this.state = {
-      isMobile: window.innerWidth < mobileBreakpoint,
-      menuOpen: false,
-      aboutModalOpen: false,
-      feedbackModalOpen: false,
       zoomLevel: DEFAULT_ZOOM,
     };
-  }
-
-  componentDidMount() {
-    window.addEventListener("resize", this.updateIsMobile);
   }
 
   componentDidUpdate(prevProps: Props) {
@@ -73,10 +55,6 @@ class MapView extends Component<Props, State> {
     }
   }
 
-  componentWillUnmount() {
-    window.removeEventListener("resize", this.updateIsMobile);
-  }
-
   get leafletElement() {
     const { mapRef } = this.props;
 
@@ -86,12 +64,6 @@ class MapView extends Component<Props, State> {
   handleZoom = () => {
     this.setState({
       zoomLevel: this.leafletElement?.getZoom() || DEFAULT_ZOOM,
-    });
-  };
-
-  updateIsMobile = () => {
-    this.setState({
-      isMobile: window.innerWidth < mobileBreakpoint,
     });
   };
 
@@ -115,69 +87,24 @@ class MapView extends Component<Props, State> {
     setLocation((latLngToArray(event.latlng) as any) as [number, number]);
   };
 
-  toggleMenu = () => {
-    const { menuOpen } = this.state;
-
-    if (menuOpen) {
-      this.setState({
-        menuOpen: false,
-      });
-    } else {
-      this.setState({
-        menuOpen: true,
-      });
-    }
-  };
-
   setView = (coordinates: [number, number]) => {
     this.leafletElement?.setView(coordinates);
   };
 
-  openAboutModal = () => {
-    this.setState({
-      aboutModalOpen: true,
-    });
-  };
-
-  closeAboutModal = () => {
-    this.setState({
-      aboutModalOpen: false,
-    });
-  };
-
-  openFeedbackModal = () => {
-    this.setState({
-      feedbackModalOpen: true,
-    });
-  };
-
-  closeFeedbackModal = () => {
-    this.setState({
-      feedbackModalOpen: false,
-    });
-  };
-
   render() {
-    const {
-      position,
-      selectedUnit,
-      units,
-      activeLanguage,
-      openUnit,
-      t,
-      mapRef,
-    } = this.props;
+    const { position, selectedUnit, units, openUnit, mapRef, t } = this.props;
 
-    const {
-      isMobile,
-      zoomLevel,
-      menuOpen,
-      aboutModalOpen,
-      feedbackModalOpen,
-    } = this.state;
+    const { zoomLevel } = this.state;
 
     return (
-      <View id="map-view" className="map-view">
+      <View
+        id="map-view"
+        className="map-view"
+        // Hide tha map from screen readers
+        aria-label={t("MAP.ACCESSIBILITY_NOTICE")}
+        aria-hidden="false"
+        tabIndex={-1}
+      >
         <Map
           ref={mapRef}
           zoomControl={false}
@@ -202,43 +129,19 @@ class MapView extends Component<Props, State> {
             selectedUnit={selectedUnit}
             openUnit={openUnit}
           />
-          <ZoomControl position="bottomright" />
+          <ZoomControl
+            position="bottomright"
+            zoomInTitle={t("MAP.ZOOM_IN")}
+            zoomOutTitle={t("MAP.ZOOM_OUT")}
+          />
           <Control
             handleClick={this.locateUser}
             className="leaflet-control-locate"
             position="bottomright"
           >
-            <OSMIcon icon="locate" />
+            <OSMIcon icon="locate" aria-label={t("MAP.LOCATE_USER")} />
           </Control>
-          <MapDropdownControl
-            id="info-dropdown"
-            handleClick={this.toggleMenu}
-            className="leaflet-control-info"
-            position={isMobile ? "bottomleft" : "topright"}
-            aria-haspopup="true"
-            aria-expanded={menuOpen}
-            activeLanguage={activeLanguage}
-            isOpen={menuOpen}
-            options={
-              <InfoMenu
-                t={t}
-                openAboutModal={this.openAboutModal}
-                openFeedbackModal={this.openFeedbackModal}
-              />
-            }
-            wrapperAttrs={{
-              role: "contentinfo",
-            }}
-          >
-            <SMIcon icon="info" aria-label={t("APP.ABOUT")} />
-          </MapDropdownControl>
         </Map>
-        {aboutModalOpen ? (
-          <AboutModal closeModal={this.closeAboutModal} t={t} />
-        ) : null}
-        {feedbackModalOpen ? (
-          <FeedbackModal closeModal={this.closeFeedbackModal} />
-        ) : null}
       </View>
     );
   }
@@ -247,76 +150,3 @@ class MapView extends Component<Props, State> {
 export default withTranslation(undefined, {
   withRef: true,
 })(MapView);
-
-type InfoMenuProps = {
-  openAboutModal: () => void;
-  openFeedbackModal: () => void;
-  t: (arg0: string) => string;
-};
-
-function InfoMenu({ openAboutModal, openFeedbackModal, t }: InfoMenuProps) {
-  return (
-    <TranslationProvider>
-      <div className="info-menu">
-        <InfoMenuItem icon="info" handleClick={openFeedbackModal}>
-          {t("MAP.INFO_MENU.GIVE_FEEDBACK")}
-        </InfoMenuItem>
-        <InfoMenuItem icon="info" handleClick={openAboutModal}>
-          {t("MAP.INFO_MENU.ABOUT_SERVICE")}
-        </InfoMenuItem>
-        <OutboundLink
-          className="info-menu-item"
-          href="http://osm.org/copyright"
-        >
-          {`\u00a9 ${t("MAP.ATTRIBUTION")} `}
-        </OutboundLink>
-      </div>
-    </TranslationProvider>
-  );
-}
-
-type InfoMenuItemProps = {
-  children: ReactNode;
-  handleClick: () => void;
-  icon?: string;
-};
-
-function InfoMenuItem({ children, handleClick, icon }: InfoMenuItemProps) {
-  return (
-    <button type="button" className="info-menu-item" onClick={handleClick}>
-      {icon ? (
-        <SMIcon
-          icon={icon}
-          style={{
-            paddingRight: 2,
-          }}
-          aria-hidden="true"
-        />
-      ) : null}
-      {children}
-    </button>
-  );
-}
-
-type AboutModalProps = {
-  closeModal: () => void;
-  t: (arg0: string) => string;
-};
-
-function AboutModal({ closeModal, t }: AboutModalProps) {
-  return (
-    <div className="about-modal-backdrop">
-      <div className="about-modal-box">
-        <div className="about-modal-controls">
-          <SMIcon icon="close" onClick={() => closeModal()} />
-        </div>
-        <div
-          className="about-modal-content" // eslint-disable-next-line react/no-danger
-          dangerouslySetInnerHTML={{
-            __html: t("MAP.ABOUT"),
-          }}
-        />
-      </div>
-    </div>
-  );
-}